refactor(collection-settings): moved concepts DTO of API in package
