+++ src/main/java/fr/siamois/domain/services/form/FormConfigService.java
@@ -14,9 +14,8 @@ import fr.siamois.domain.services.vocabulary.ConceptService;
 import fr.siamois.domain.services.vocabulary.VocabularyService;
 import fr.siamois.dto.entity.vocabulary.ConceptCollectionDTO;
 import fr.siamois.dto.entity.vocabulary.ConceptDTO;
-import fr.siamois.dto.entity.vocabulary.VocabularyDTO;
 import fr.siamois.infrastructure.api.ConceptApi;
-import fr.siamois.infrastructure.api.dto.ConceptBranchDTO;
+import fr.siamois.infrastructure.api.dto.concept.ConceptBranchDTO;
 import fr.siamois.infrastructure.database.repositories.form.config.FieldFormConfigRepository;
 import fr.siamois.mapper.vocabulary.VocabularyMapper;
 import fr.siamois.utils.vocabulary.ConceptApiUtils;
