+++ src/main/java/fr/siamois/infrastructure/api/dto/concept/ConceptRemoteAutocompleteDTO.java
@@ -1,4 +1,4 @@
-package fr.siamois.infrastructure.api.dto;
+package fr.siamois.infrastructure.api.dto.concept;
 
 public record ConceptRemoteAutocompleteDTO(Long identifier, String uri, String label, Boolean isAltLabel,
                                            String definition) {
